Move `shouldInlineLogicalExpression` to binaryish.js
